refactor: move common func to separate file
