+++ internal/apiserver/database/kolide_issues.go
@@ -3,7 +3,6 @@ package database
 import (
 	"context"
 	"database/sql"
-	"errors"
 	"fmt"
 	"strings"
 	"time"
@@ -15,60 +14,6 @@ import (
 	"github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
-
-func (db *database) getDeviceIssues(ctx context.Context, device *sqlc.Device) ([]*pb.DeviceIssue, error) {
-	if !db.kolideEnabled {
-		return nil, nil
-	}
-
-	if device.ExternalID.String == "" {
-		return []*pb.DeviceIssue{{
-			Title:         "No Kolide device ID found for device",
-			Message:       "Make sure you've installed Kolide according to the documentation at https://doc.nais.io/operate/naisdevice/how-to/install-kolide",
-			Severity:      pb.Severity_Critical,
-			DetectedAt:    timestamppb.Now(),
-			LastUpdated:   timestamppb.Now(),
-			ResolveBefore: timestamppb.New(time.Time{}),
-		}}, nil
-	}
-
-	lastSeen := stringToTime(device.LastSeen.String)
-	lastUpdated := stringToTime(device.LastUpdated.String)
-
-	issues, err := db.queries.GetKolideIssuesForDevice(ctx, device.ExternalID.String)
-	if err != nil && !errors.Is(err, sql.ErrNoRows) {
-		return nil, fmt.Errorf("getting kolide issues: %w", err)
-	}
-
-	deviceIssues := make([]*pb.DeviceIssue, 0)
-	if lastSeenIssue := generateLastSeenIssue(lastSeen, lastUpdated); lastSeenIssue != nil {
-		deviceIssues = append(deviceIssues, lastSeenIssue)
-	}
-
-	for _, issue := range issues {
-		checkTags := strings.Split(issue.Tags, ",")
-		checkSeverity := kolideCheckSeverity(checkTags, db.log)
-
-		if checkSeverity == pb.Severity_Info || issue.ResolvedAt.String != "" || issue.Ignored {
-			continue
-		}
-
-		issueDetectedAt := stringToTime(issue.DetectedAt)
-		issueLastUpdated := stringToTime(issue.LastUpdated)
-		issueResolveBefore := issueDetectedAt.Add(graceTime(checkSeverity))
-
-		deviceIssues = append(deviceIssues, &pb.DeviceIssue{
-			Title:         issue.Title,
-			Message:       issue.Description,
-			Severity:      checkSeverity,
-			DetectedAt:    timestamppb.New(issueDetectedAt),
-			LastUpdated:   timestamppb.New(issueLastUpdated),
-			ResolveBefore: timestamppb.New(issueResolveBefore),
-		})
-	}
-
-	return deviceIssues, nil
-}
 
 const (
 	lastSeenGracePeriod = 48 * time.Hour
@@ -112,15 +57,6 @@ If it's not and you don't know why - re-install the launcher by asking @Kolide f
 	return template(msg, lastUpdated)
 }
 
-// Devices are allowed to connect this long after a failed check is triggered.
-const (
-	DurationCritical = 0
-	DurationDanger   = time.Hour
-	DurationWarning  = time.Hour * 24 * 2
-	DurationNotice   = time.Hour * 24 * 7
-	DurationUnknown  = time.Hour * 24 * 30
-)
-
 func kolideCheckSeverity(tags []string, log logrus.FieldLogger) pb.Severity {
 	highest := pb.Severity(-1)
 
@@ -152,21 +88,6 @@ func kolideCheckSeverity(tags []string, log logrus.FieldLogger) pb.Severity {
 	}
 
 	return highest
-}
-
-func graceTime(severity pb.Severity) time.Duration {
-	switch severity {
-	case pb.Severity_Notice:
-		return DurationNotice
-	case pb.Severity_Warning:
-		return DurationWarning
-	case pb.Severity_Danger:
-		return DurationDanger
-	case pb.Severity_Critical:
-		return DurationCritical
-	default:
-		return DurationUnknown
-	}
 }
 
 func (db *database) UpdateKolideChecks(ctx context.Context, checks []*kolide.Check) error {
